Refactor command packages according to actual scoping
