+++ src/main/java/seedu/address/logic/commands/global/ClearCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.catalogue;
+package seedu.address.logic.commands.global;
 
 import static java.util.Objects.requireNonNull;
