+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -13,7 +13,7 @@ import java.util.HashMap;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.catalogue.AddCommand;
+import seedu.address.logic.commands.global.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.meeting.Meeting;
 import seedu.address.model.project.Deadline;
