+++ src/main/java/seedu/address/logic/parser/EditTaskCommandParser.java
@@ -14,7 +14,7 @@ import java.util.Optional;
 import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.catalogue.EditCommand;
+import seedu.address.logic.commands.global.EditCommand;
 import seedu.address.logic.commands.project.EditTaskCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.task.Task;
