+++ src/main/java/seedu/address/logic/parser/MainCatalogueParser.java
@@ -7,13 +7,13 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.catalogue.AddCommand;
-import seedu.address.logic.commands.catalogue.ClearCommand;
-import seedu.address.logic.commands.catalogue.DeleteCommand;
-import seedu.address.logic.commands.catalogue.EditCommand;
-import seedu.address.logic.commands.catalogue.FindCommand;
-import seedu.address.logic.commands.catalogue.ListCommand;
-import seedu.address.logic.commands.catalogue.StartCommand;
+import seedu.address.logic.commands.global.AddCommand;
+import seedu.address.logic.commands.global.ClearCommand;
+import seedu.address.logic.commands.global.DeleteCommand;
+import seedu.address.logic.commands.global.EditCommand;
+import seedu.address.logic.commands.global.FindCommand;
+import seedu.address.logic.commands.global.ListCommand;
+import seedu.address.logic.commands.global.StartCommand;
 import seedu.address.logic.commands.global.ExitCommand;
 import seedu.address.logic.commands.global.HelpCommand;
 import seedu.address.logic.commands.project.AddTaskCommand;
