+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -18,8 +18,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.catalogue.AddCommand;
-import seedu.address.logic.commands.catalogue.ListCommand;
+import seedu.address.logic.commands.global.AddCommand;
+import seedu.address.logic.commands.global.ListCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
