+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -15,7 +15,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.catalogue.EditCommand;
+import seedu.address.logic.commands.global.EditCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.MainCatalogue;
 import seedu.address.model.Model;
