+++ src/test/java/seedu/address/logic/commands/global/AddCommandTest.java
@@ -1,9 +1,10 @@
-package seedu.address.logic.commands.catalogue;
+package seedu.address.logic.commands.global;
 
 import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import seedu.address.logic.commands.global.AddCommand;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import java.nio.file.Path;
