+++ src/test/java/seedu/address/logic/commands/global/EditCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.catalogue;
+package seedu.address.logic.commands.global;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -10,6 +10,8 @@ import static seedu.address.logic.commands.CommandTestUtil.VALID_PROJECT_TAG_HAN
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.CommandTestUtil.showProjectAtIndex;
+import seedu.address.logic.commands.global.ClearCommand;
+import seedu.address.logic.commands.global.EditCommand;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PROJECT;
 import static seedu.address.testutil.TypicalIndexes.INDEX_SECOND_PROJECT;
 import static seedu.address.testutil.TypicalProjects.getTypicalMainCatalogue;
@@ -18,7 +20,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.catalogue.EditCommand.EditProjectDescriptor;
+import seedu.address.logic.commands.global.EditCommand.EditProjectDescriptor;
 import seedu.address.model.MainCatalogue;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
