+++ src/test/java/seedu/address/logic/commands/global/FindCommandTest.java
@@ -1,10 +1,11 @@
-package seedu.address.logic.commands.catalogue;
+package seedu.address.logic.commands.global;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.commons.core.Messages.MESSAGE_PROJECTS_LISTED_OVERVIEW;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
+import seedu.address.logic.commands.global.FindCommand;
 import static seedu.address.testutil.TypicalProjects.CARACTIVE;
 import static seedu.address.testutil.TypicalProjects.ELFIE;
 import static seedu.address.testutil.TypicalProjects.FANS;
