+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -35,7 +35,7 @@ import static seedu.address.testutil.TypicalProjects.BOT;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.catalogue.AddCommand;
+import seedu.address.logic.commands.global.AddCommand;
 import seedu.address.model.project.Deadline;
 import seedu.address.model.project.Project;
 import seedu.address.model.project.ProjectDescription;
