+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -8,7 +8,7 @@ import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.catalogue.FindCommand;
+import seedu.address.logic.commands.global.FindCommand;
 import seedu.address.model.project.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
