+++ src/test/java/seedu/address/testutil/EditProjectDescriptorBuilder.java
@@ -4,7 +4,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.catalogue.EditCommand.EditProjectDescriptor;
+import seedu.address.logic.commands.global.EditCommand.EditProjectDescriptor;
 import seedu.address.model.project.Deadline;
 import seedu.address.model.project.Project;
 import seedu.address.model.project.ProjectDescription;
