+++ src/test/java/seedu/address/testutil/ProjectUtil.java
@@ -9,8 +9,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TASK;
 
 import java.util.Set;
 
-import seedu.address.logic.commands.catalogue.AddCommand;
-import seedu.address.logic.commands.catalogue.EditCommand.EditProjectDescriptor;
+import seedu.address.logic.commands.global.AddCommand;
+import seedu.address.logic.commands.global.EditCommand.EditProjectDescriptor;
 import seedu.address.model.project.Project;
 import seedu.address.model.tag.ProjectTag;
 import seedu.address.model.task.Task;
